Move items to types folder
