+++ src/blockchain/NEAR.ts
@@ -1,82 +1,6 @@
 import { connect, Contract, keyStores, WalletConnection } from 'near-api-js'
 import { getBlockchainType } from '../utils'
 
-export interface Three0Contract extends Contract {
-	valid_database: ValidDatabaseFunction
-	get_user: GetUserFunction
-	user_action: UserActionFunction
-	get_storage: GetStorageFunction
-	set_nonce: SetNonceFunction
-}
-
-type ValidDatabaseFunction = (
-	database_id: ValidDatabaseParams
-) => Promise<boolean>
-type GetUserFunction = (user_id: GetUserParams) => Promise<User>
-type UserActionFunction = (action: UserActionParams) => Promise<void>
-type GetStorageFunction = () => Promise<string>
-type SetNonceFunction = () => Promise<number>
-
-interface User {
-	account_id: string
-	is_online: boolean
-	created_at: bigint
-	last_online: bigint
-}
-
-interface ValidDatabaseParams {
-	address: string
-}
-
-interface GetUserParams {
-	account_id: string
-}
-
-interface UserActionParams {
-	action: UserActionType
-}
-
-export enum UserActionType {
-	LOGIN = 'LOGIN',
-	LOGOUT = 'LOGOUT',
-}
-
-export interface StorageContract extends Contract {
-	list_files: ListFilesFunction
-	get_file: GetFileFunction
-	nft_mint: NftMintFunction
-}
-
-type ListFilesFunction = (params: ListFilesParams) => Promise<string[]>
-type GetFileFunction = (params: GetFileParams) => Promise<FileMetadata>
-type NftMintFunction = (
-	params: NftMintParams,
-	attached_gas: string,
-	attached_deposit: string
-) => Promise<void>
-
-interface ListFilesParams {
-	path: string
-}
-interface GetFileParams {
-	file_path: string
-}
-interface NftMintParams {
-	token_id: string
-	metadata: FileMetadata
-	path: string
-	receiver_id: string
-}
-
-export interface FileMetadata {
-	title: string
-	description: string
-	media: string
-	media_hash: string
-	file_type: string
-	issued_at: number
-}
-
 export function getNearConfig() {
 	const CONTRACT_NAME = globalThis.projectConfig.contractName
 	const chainType = getBlockchainType()
