+++ src/types/auth.d.ts
@@ -1,1 +1,2 @@
 declare module 'localstorage-memory'
+declare module 'near-seed-phrase'
