+++ src/types/config.ts
@@ -1,5 +1,7 @@
+// declare module '@three0dev/js-sdk/types' {
 export interface ProjectConfig {
 	chainType: string
 	contractName: string
 	projectId: string
 }
+// }
