--- src/types/global.d.ts
@@ -1,18 +1,0 @@
-/* eslint-disable no-var */
-/* eslint-disable vars-on-top */
-import { WalletConnection } from 'near-api-js'
-import OrbitDB from 'orbit-db'
-import { ProjectConfig } from './config'
-import { Three0Contract, StorageContract } from '../blockchain/NEAR'
-
-// TODO: Account for Contract Object
-declare global {
-	var projectConfig: ProjectConfig
-	var orbitdb: OrbitDB
-	var walletConnection: WalletConnection
-	var accountId: string
-	var contract: Three0Contract
-	var storageContract: StorageContract
-}
-
-export {}
